+++ packages/react/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.0.0-next.28
 
 - add `paneDistanceClick` prop (max distance between mousedown/up that will trigger a click)
+- returned nodes in `onNodeDragStop` are set to `dragging=false` 
 
 ## 12.0.0-next.27